requirements.txt
streamlit>=1.37
python-dotenv>=1.0
pandas>=2.2
plotly>=5.22
pydantic>=2.8

watchdog>=3.0
requests>=2.32
python-dotenv>=1.0
